requirements.txt
streamlit
langchain
langchain-anthropic
langchain-community
langchain-huggingface
langchain-text-splitters
chromadb
sentence-transformers
pypdf
python-dotenv 